Back-end/requirementorg.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.6.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.0
#cliapp==1.20180812.1
click==8.1.7
cloud-init==23.3.1
#cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.6.0
Flask==3.0.0
frozenlist==1.4.0
gunicorn==21.2.0
#gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
Markdown==3.3.6
MarkupSafe==2.1.3
more-itertools==8.10.0
multidict==6.0.4
#mysqlclient==2.2.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-linux-procfs==0.6.3
python-magic==0.4.24
pytz==2022.1
pyudev==0.22.0
PyYAML==5.4.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.5.6
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
#ttystatus==0.38
Twisted==22.1.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
#virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==3.0.0
yarl==1.9.2
zeep==4.2.1
zipp==1.0.0
zope.interface==5.4.0
